Remove the annoying warning: onnly the maptplot lib warning remains https://github.com/matplotlib/matplotlib/issues/5209.

+++ generateBeam.py
@@ -236,7 +236,7 @@ else:
 highres_phase = np.linspace(-180,180,10*res)
 resampled = np.zeros((int(nch),int(10*res)))
 for nfr in range(len(freq)):
-    resampled[nfr] = sci_sig.resample(profile[nfr], 10*res)
+    resampled[nfr] = sci_sig.resample(profile[nfr], int(10*res))
 
 if all(i > 10 for i in SN):
     average_profile = []
